hw11/requirements.txt
fastapi==0.118.2
psycopg==3.2.10
psycopg-binary==3.2.10
uvicorn==0.37.0
faker==37.11.0
pyjwt==2.10.1
redis==6.4.0
couchbase==4.5.0
ulid-py==1.1.0
websockets==15.0.1
aio-pika==9.5.7
grpcio==1.76.0
protobuf==6.33.0
prometheus-fastapi-instrumentator==7.1.0